pyproject.toml
[project]
name = "ntk-vk-first-sem"
version = "0.1.0"
description = ""
authors = [
    {name = "MaksimKislyakov",email = "maksimkislakov35@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-telegram-bot (>=22.5,<23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
